shank reorg
