+++ solores/src/idl_format/mod.rs
@@ -36,6 +36,7 @@ pub trait IdlFormat {
 macro_rules! gen_body_newtype_slice {
     () => {
         fn gen_body(&self) -> TokenStream {
+            use quote::ToTokens;
             let elems = self.0.iter().map(|e| e.into_token_stream());
             let mut res = quote! {};
             res.extend(elems);
