+++ solores/src/idl_format/shank/errors/error.rs
